Creacion de la interfaz de agentes #24
Se ha extraido la mayoria de la funcionalidad del script principal, creando la interfaz BaseAgent y la clase LocalAgent que hereda de esta e implementa todos los metodos necesarios para realizar las diferentes acciones para el funcionamiento del agente, como la creacion del modelo, el entrenamiento, etc.
También se han organizado los ficheros en carpetas, creando una para los módulos, y otra para las clases.

+++ Agente/custom_env/redes_opticas_env.py
@@ -15,8 +15,8 @@ class RedesOpticasEnv(gym.Env):
         self.velocidadContratada = vt_contratada
         self.Max_bits_ONT=self.velocidadContratada*self.temp_ciclo
 
-        self.observation_space = spaces.Box(low=0, high=self.Max_bits_ONT, shape=(self.num_ont,), dtype=np.float32)
-        self.action_space = spaces.Box(low=-self.Max_bits_ONT, high=self.Max_bits_ONT, shape=(self.num_ont,), dtype=np.float32)
+        self.observation_space = spaces.Box(low=0, high=self.Max_bits_ONT, shape=(self.num_ont,), dtype=np.float64)
+        self.action_space = spaces.Box(low=-self.Max_bits_ONT, high=self.Max_bits_ONT, shape=(self.num_ont,), dtype=np.float64)
 
         self.step_durations = []
         self.trafico_entrada = []
